+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v2.0.42 ~ 
 
+  * Trigs128: NEW MODULE!
   * Trigs: added 'Repeat First Row Down' menu item
   * NoteSeq16: added 'Repeat 16 Steps'
   * NoteSeq: added 'Repeat 16 Steps'